requirements.txt
python-dotenv
openai
openai-agents
requests
pypdf
gradio
fastapi
uvicorn[standard]
python-multipart
httpx
tiktoken
pinecone
